refactor: compose recipes client-side only; drop server compose + endpoint refs
Recipe composition now lives entirely in the UI. The endpoint is written as a
plain, concrete spec (model/engine/resources/variables/env) — it no longer
carries recipe refs, and the controller no longer composes at reconcile time.

- Remove expandCatalogRef from the endpoint controller and delete the server
  compose engine (internal/recipe/compose.go, normalize.go). This also removes
  the whole expand/reconcile path that caused the model-less infinite-reconcile
  risk, and ends the Go/TS dual-compose maintenance burden.
- Keep recipe validation: the shared per-feature validation helpers move into
  validate.go, which the POST /model_catalogs middleware still uses to reject
  malformed recipe catalogs at the data-entry boundary.
- Drop the endpoint_spec recipe columns (model_catalog/variant/feature_selections)
  by editing the unmerged migration 073 in place; keep the model_catalog_spec
  base/variants/features columns. Revert the dbtest endpoint_spec ROWs to 8
  fields and remove the FeatureSelection type + api_gen fields.
