+++ db/dbtest/validation_test.go
@@ -666,9 +666,6 @@ func TestEndpointAcceleratorValidation(t *testing.T) {
 					ROW(1)::api.replica_spec,
 					NULL,
 					NULL,
-					NULL,
-					NULL,
-					NULL,
 					NULL
 				)::api.endpoint_spec,
 				ROW('test-ep-accel', NULL, 'test-workspace', NULL, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP, '{}'::json, '{}'::json)::api.metadata
